+++ example/01_hello/program.c
@@ -11,11 +11,11 @@ void draw_star(int x, int y)
 
 int main(int argc, char* argv[])
 {
-    VGS_VREG_BMP0 = 1; // BG0: Bitmap Mode
-    VGS_VREG_BMP1 = 1; // BG1: Bitmap Mode
-    VGS_VREG_BMP2 = 1; // BG2: Bitmap Mode
-    VGS_VREG_BMP3 = 0; // BG3: Character Pattern Mode
-    VGS_VREG_SPOS = 3; // Sprite: Displayed on top of BG3
+    vgs_draw_mode(0, TRUE);  // BG0: Bitmap Mode
+    vgs_draw_mode(1, TRUE);  // BG1: Bitmap Mode
+    vgs_draw_mode(2, TRUE);  // BG2: Bitmap Mode
+    vgs_draw_mode(3, FALSE); // BG3: Character Pattern Mode
+    VGS_VREG_SPOS = 3;       // Sprite: Displayed on top of BG3
     vgs_print_bg(3, 11, 12, 0, "HELLO VGS-X WORLD!");
 
     uint32_t col = 1;
@@ -51,17 +51,17 @@ int main(int argc, char* argv[])
     vgs_sprite(0, TRUE, 20, 90, 2, 0, '1');
     vgs_draw_box(2, VRAM_WIDTH - 20 - 32, 86, VRAM_WIDTH - 20 - 32 + 31, 86 + 31, 0xF02020);
     vgs_sprite(1, TRUE, VRAM_WIDTH - 20 - 32, 86, 3, 0, 'A');
-    OAM[1].scale = 10;
+    vgs_oam(1)->scale = 10;
 
     vgs_sprite(2, TRUE, VRAM_WIDTH / 2 - 32 - 8, 20, 0, 0, 'V');
     vgs_sprite(3, TRUE, VRAM_WIDTH / 2 - 4, 20, 0, 0, 'G');
     vgs_sprite(4, TRUE, VRAM_WIDTH / 2 + 32, 20, 0, 0, 'S');
-    OAM[2].scale = 300;
-    OAM[2].rotate = 90;
-    OAM[3].scale = 200;
-    OAM[3].rotate = 120;
-    OAM[4].scale = 100;
-    OAM[4].rotate = 150;
+    vgs_oam(2)->scale = 300;
+    vgs_oam(2)->rotate = 90;
+    vgs_oam(3)->scale = 200;
+    vgs_oam(3)->rotate = 120;
+    vgs_oam(4)->scale = 100;
+    vgs_oam(4)->rotate = 150;
     int sa[4] = {10, 10, 10, 10};
 
     const int32_t baseX = 120;
@@ -83,31 +83,31 @@ int main(int argc, char* argv[])
         VGS_VREG_SY1 = 1;
         vgs_draw_pixel(1, vgs_rand() % VRAM_WIDTH, 0, vgs_rand32());
         // vgs_draw_character(1, vgs_rand() % 320 - 4, 0, FALSE, 0, 0x20 + (vgs_rand() & 0x3F));
-        OAM[0].rotate += 3;
+        vgs_oam(0)->rotate += 3;
 
         for (int i = 0; i < 4; i++) {
-            OAM[i + 1].scale += sa[i];
-            if (400 < OAM[i + 1].scale && 0 < sa[i]) {
+            vgs_oam(i + 1)->scale += sa[i];
+            if (400 < vgs_oam(i + 1)->scale && 0 < sa[i]) {
                 sa[i] = -10;
                 if (i == 0) {
                     vgs_sfx_play(0);
                 }
-            } else if (OAM[i + 1].scale < 20 && sa[i] < 0) {
+            } else if (vgs_oam(i + 1)->scale < 20 && sa[i] < 0) {
                 sa[i] = 10;
             }
             if (i) {
-                OAM[i + 1].rotate += i * 3;
+                vgs_oam(i + 1)->rotate += i * 3;
             }
         }
-        OAM[6].visible = VGS_KEY_UP;
-        OAM[7].visible = VGS_KEY_DOWN;
-        OAM[8].visible = VGS_KEY_LEFT;
-        OAM[9].visible = VGS_KEY_RIGHT;
-        OAM[10].visible = VGS_KEY_START;
-        OAM[11].visible = VGS_KEY_A;
-        OAM[12].visible = VGS_KEY_B;
-        OAM[13].visible = VGS_KEY_X;
-        OAM[14].visible = VGS_KEY_Y;
+        vgs_oam(6)->visible = VGS_KEY_UP;
+        vgs_oam(7)->visible = VGS_KEY_DOWN;
+        vgs_oam(8)->visible = VGS_KEY_LEFT;
+        vgs_oam(9)->visible = VGS_KEY_RIGHT;
+        vgs_oam(10)->visible = VGS_KEY_START;
+        vgs_oam(11)->visible = VGS_KEY_A;
+        vgs_oam(12)->visible = VGS_KEY_B;
+        vgs_oam(13)->visible = VGS_KEY_X;
+        vgs_oam(14)->visible = VGS_KEY_Y;
     }
 
     return 0;