+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Version 0.9.0
 
 - CRT: Added a sprite function: `vgs_oam`
+- CRT: Added a bitmap function: `vgs_draw_mode`
 - CRT: Renamed the OAM structure to ObjectAttributeMemory to avoid name conflicts with `#define` literals. **(Disruptive)**
 - Toolchain: Supports 16-bit format output with `bin2var`.
 - Toolchain: Added `csv2var`, a tool to convert Tiled Map Editor CSV files.